Add minigames slash command (#3631)
