+++ src/tasks/minions/minigames/sepulchreActivity.ts
@@ -67,6 +67,15 @@ export default class extends Task {
 				previousCL
 			);
 
-		handleTripFinish(this.client, user, channelID, str, ['sepulchre', [], true], image!, data, itemsAdded);
+		handleTripFinish(
+			this.client,
+			user,
+			channelID,
+			str,
+			['minigames', { sepulchre: { start: {} } }, true],
+			image!,
+			data,
+			itemsAdded
+		);
 	}
 }
